% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comediator_region.R
\name{comediator_region}
\alias{comediator_region}
\title{Create list with mediators in region}
\usage{
comediator_region(
  pheno_name,
  chr_id,
  scan_window,
  covar,
  analyses_tbl,
  peaks,
  qtls = 2,
  pmap,
  pheno_data
)
}
\arguments{
\item{pheno_name}{phenotype name}

\item{chr_id, scan_window}{chromosome and start and end value}

\item{covar}{covariate data frame}

\item{analyses_tbl}{table of analyses}

\item{peaks}{table of peaks}

\item{qtls}{number of drivers (1 or 2; default is 2)}

\item{pmap}{physical map}

\item{pheno_data}{phenotype data}
}
\description{
Create list with mediators in region
}
