% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comediator_region.R
\name{comediator_type}
\alias{comediator_type}
\title{Get mediators of same type as pheno_name}
\usage{
comediator_type(comed, peaks, pheno_name, doThis)
}
\description{
Get mediators of same type as pheno_name
}
