% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQtl2.R
\name{fitQtl2}
\alias{fitQtl2}
\title{Fit target relative to driver using qtl2 fit1}
\usage{
fitQtl2(driver, target, kinship = NULL, sdp, allele, ...)
}
\description{
Fit a model for a target and get detailed results
about estimated coefficients and individuals contributions to the LOD score.
}
