% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_triad_qtl2.R
\name{mediation_triad_qtl2}
\alias{mediation_triad_qtl2}
\title{Test mediation across set of indexed drivers for `qtl2` data}
\usage{
mediation_triad_qtl2(
  target,
  mediator,
  annotation,
  covar_tar = NULL,
  covar_med = NULL,
  genoprobs,
  map,
  chr,
  pos,
  kinship = NULL,
  sdp = NULL,
  ...
)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{genoprobs}{genoprob object of class [qtl2::calc_genoprob()]}

\item{map}{list of map positions}

\item{kinship}{optional kinship matrix among individuals}

\item{...}{additional parameters for [mediation_index()]}

\item{drop_lod}{drop in `LOD` (= `LR/log(10)`) to define index set}

\item{min_lod}{minimum lod to consider (default \code{3})}

\item{query_variant}{function to query variant database}

\item{cores}{number of cores to use}

\item{target_scan}{optional object from [qtl2::scan1snps()] for target (created if missing)}
}
\value{
Object of class `mediation_qtl2`, which inherits from class [mediation_index()]
}
\description{
Test mediation across set of indexed drivers for `qtl2` data
}
\examples{
dirpath <- "https://raw.githubusercontent.com/rqtl/qtl2data/master/DOex"

# Read DOex example cross from 'qtl2data'
DOex <- subset(qtl2::read_cross2(file.path(dirpath, "DOex.zip")), chr = "2")

# Download genotype probabilities
tmpfile <- tempfile()
download.file(file.path(dirpath, "DOex_genoprobs_2.rds"), tmpfile, quiet=TRUE)
pr <- readRDS(tmpfile)
unlink(tmpfile)

}
