% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_sdp.R
\name{triad_sdp}
\alias{triad_sdp}
\title{Locate triad for mediation}
\usage{
triad_sdp(
  chr,
  pos,
  scan_window,
  pheno_name,
  pheno_data,
  genoprobs,
  map,
  analyses_tbl,
  kinship = NULL,
  snpinfo = query_variants(chr, scan_window[1], scan_window[2]),
  ...
)
}
\arguments{
\item{chr, pos, scan_window}{locators for region}

\item{pheno_name, pheno_data}{name and values of phenotype data}

\item{genoprobs}{genotype probabilities}

\item{map}{map of genome}

\item{analyses_tbl}{analyses table}

\item{kinship}{kinship information (optional)}

\item{...}{additional arguments}

\item{snpinf}{SNP information data frame (from user supplied \code{query_variants})}
}
\description{
Given chr, pos and scan_window, target name and snpinfo,
find sdp for best pattern.
}
